intel_realsense_payload_for_spotsim2real/requirements.txt
bosdyn-api==3.2.3
bosdyn-choreography-client==3.2.3
bosdyn-choreography-protos==3.2.3
bosdyn-client==3.2.3
bosdyn-core==3.2.3
bosdyn-mission==3.2.3
numpy
opencv-python
pyzmq
pyrealsense2